Add: header connect redux store
