+++ src/App.js
@@ -22,30 +22,23 @@ class App extends Component {
         }).isRequired,
         getToken: PropTypes.func.isRequired,
         fetchCurrentUser: PropTypes.func.isRequired,
-        user: PropTypes.shape({
-            isFetching: PropTypes.bool,
-        }).isRequired,
     };
 
     async componentDidMount() {
         await this.props.getToken();
         if (this.props.auth.isAuthenticated) {
+            debugger;
             this.props.fetchCurrentUser(this.props.auth.token);
         }
     }
 
     render() {
-        const { isAuthenticated } = this.props.auth;
-        const { isLoading } = this.props.auth;
-        const { error } = this.props.auth;
+        const { isAuthenticated, isLoading, error } = this.props.auth;
 
         if (isLoading) {
             return (<p>Loading!</p>);
         }
 
-        if (this.props.user.isFetching || this.props.user.isFetching === null) {
-            return <p>Loading!</p>;
-        }
 
         if (error !== null) {
             return (
@@ -57,11 +50,7 @@ class App extends Component {
 
         if (isAuthenticated) {
             return (
-                <Wrapper
-                    auth={isAuthenticated}
-                    userName={this.props.user.data.name}
-                    profileImg={this.props.user.data.profileImg}
-                >
+                <Wrapper>
                     <Switch>
                         <Route path="/home" component={HomePage} />
                         <Route path="/event" component={EventDetails} />
@@ -72,7 +61,7 @@ class App extends Component {
         }
 
         return (
-            <Wrapper auth={isAuthenticated}>
+            <Wrapper>
                 <Switch>
                     <Route path="/" exact component={HomePage} />
                     <Route path="/home" component={HomePage} />
@@ -85,10 +74,9 @@ class App extends Component {
     }
 }
 
-const mapStateToProps = ({ auth, user }) => {
+const mapStateToProps = ({ auth }) => {
     return {
         auth,
-        user,
     };
 };
