+++ src/common/Wrapper.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 
 import Header from './Header';
 import Footer from './Footer';
@@ -7,19 +6,11 @@ import Footer from './Footer';
 const Wrapper = (props) => {
     return (
         <div>
-            <Header
-                isAuthenticated={props.auth}
-                userName={props.userName}
-                profileImg={props.profileImg}
-            />
+            <Header />
             {props.children}
             <Footer />
         </div>
     );
-};
-
-Wrapper.propTypes = {
-    auth: PropTypes.bool.isRequired,
 };
 
 export default Wrapper;
